Remove DefaultAzureCredential from codebase
