+++ docs/source/reference/options.md
@@ -54,17 +54,6 @@ more details.
 - `max_operation_retries: [int]` The maximum number of retries for Azure ML operations like resource creation and download.
 The default value is 3. User can increase if there are network issues and the operations fail.
 - `operation_retry_interval: [int]` The initial interval in seconds between retries for Azure ML operations like resource creation and download. The interval doubles after each retry. The default value is 5. User can increase if there are network issues and the operations fail.
-- `default_auth_params: Dict[str, Any]` Default auth parameters for AzureML client. Please refer to [azure DefaultAzureCredential](https://learn.microsoft.com/en-us/python/api/azure-identity/azure.identity.defaultazurecredential?view=azure-python#parameters) for more details. For example, if you want to exclude managed identity credential, you can set the following:
-
-    ```json
-    "azureml_client": {
-        // ...
-        "default_auth_params": {
-            "exclude_managed_identity_credential": true
-        }
-    }
-    ```
-
 - `keyvault_name: [str]` The keyvault name to retrieve secrets.
 
 ### Example
