+++ olive/common/hf/login.py
@@ -5,6 +5,8 @@
 import logging
 import os
 
+from olive.common.utils import get_credentials
+
 logger = logging.getLogger(__name__)
 
 
@@ -17,13 +19,12 @@ def huggingface_login(token: str):
 def aml_runner_hf_login():
     hf_login = os.environ.get("HF_LOGIN")
     if hf_login:
-        from azure.identity import DefaultAzureCredential
         from azure.keyvault.secrets import SecretClient
 
         keyvault_name = os.environ.get("KEYVAULT_NAME")
         logger.debug("Getting token from keyvault %s", keyvault_name)
 
-        credential = DefaultAzureCredential()
+        credential = get_credentials()
         secret_client = SecretClient(vault_url=f"https://{keyvault_name}.vault.azure.net/", credential=credential)
         token = secret_client.get_secret("hf-token").value
         huggingface_login(token)
